Update MASTER.MSG
TMA_MSG_DONTKNOW_START (and END) are voicelines. I'm as dumb as the person who scripted the Barter modification on various FO2 NPCs under the belief that it "worked the other way around", and thus Randal lets you have about EVERYTHING... whereas celebrities in New Reno have to pay an arm and a leg (and Eldridge your whole body when he's supposed to only charge 1% of you after beating the game).